Improve memory usage and speed.
This patch is a big refactoring with a lot of different changes
in it. The main change is the use of shared_ptr instead of raw pointers.
This allow very efficient caching mechanism which reduces dramatically
the memory consumption as well as the parsing time.

Objects aren't copyable and are now moved around when necessary.

Other different improvements and cleanup have been done along the way too.
e.g.
- Documentation update
- Fix broken tests
- ...