Fix issue of dividing by zero in comparators.
When the tool would generate code for comparing structs with
array members, it uses the following formula.

sizeof(f)/sizeof(f[0])

However, if the type of f is a struct without a member,
sizeof(f[0]) == 0.

This patch remove the generation of that code because it doesn't make
sense to compare empty structs.