Support C++ mock
This page refactors the code to support C++ mock. The main point of
refactoring is the usage of std::shared_ptr to return the
FunctionDeclaration objects. This change is needed because the
FunctionDeclaration object need now to be stored in a ComposableType
object.

EasyMock now accept the `--mock-cpp` parameter to allow mocking C++
headers.