Add support to mock function with references.
Both return value as well as function parameters are supported.

The mock uses an std::reference_wrapper to be able to store and copy
the reference to be returned.

A new C++ MockedCppFunction struct has been introduce to support
the reference wrapper storage.

Some improvements on the doxygen documentation has been done.
From version 1.9.5, no error is reported in the documentation.